refactor(engine-host): move the config/model/permission web below the seam
Extract the config-resolution helpers (DEFAULT_MODEL, ModelSource,
lookup_default_model, the .scode.json loaders, model-alias expansion, and the
permission-mode resolvers) out of rusty-sudocode-cli into a new
engine_host::config module. These resolve an engine input -- which model to
talk to, which permission mode to enforce -- not a rendering concern, so they
belong below the seam where both the REPL and engine-acp build a session.

The move set is closed: resolve_model_alias and config_alias_for_current_dir
join it so engine-host's callee surface is confined to runtime:: and
engine_core::SudoCodeConfig, with no back-reference into the CLI crate (no
cycle). The renderer re-imports the symbols by their existing cli::args:: /
crate:: paths, so every call site is unchanged. ModelProvenance stays with the
renderer (it is status-display audit, not resolution).

First staged step of the engine-host extraction (#120).
