+++ rust/crates/engine-host/Cargo.toml
@@ -11,6 +11,12 @@ publish.workspace = true
 # `engine-acp` depend on it to build a session; neither the wire types nor the
 # renderer live here. Dependencies are added as the engine code moves in.
 [dependencies]
+# `SudoCodeConfig` (the config SSOT, re-exported by engine-core) for model/alias
+# resolution.
+engine-core = { path = "../engine-core" }
+# `ConfigLoader`, `PermissionMode`, `ResolvedPermissionMode`, the workspace-root
+# lookup, and the model-capability seeding the config helpers drive.
+runtime = { path = "../runtime" }
 
 [lints]
 workspace = true
